[CGAME] - Changed comment formatting.
Altered comment formatting to easier locate UnityMod changes in base files by changing named comments to "Unity". Code by other developers will be credited in the in-game credits and on the repo.

We ask that any developers looking to contribute to UnityMod follows the guidelines laid out by us to avoid conflicts and confusion.

For own Unity files, such as cg_unity.c, comments should not be prefixed:
// Description.

For comments in base files the formatting is as follows:
Inline / Single line comment:
//[Unity /] - Description.

Code block comment:
//[Unity] - Description.

//[/Unity]